+++ lib/sprockets/loader.rb
@@ -200,7 +200,7 @@ module Sprockets
         else
           dependencies << build_file_digest_uri(unloaded.filename)
           metadata = {
-            digest: file_digest(unloaded.filename),
+            digest: digest(self.version + file_digest(unloaded.filename)),
             length: self.stat(unloaded.filename).size,
             dependencies: dependencies
           }