+++ test/test_environment.rb
@@ -720,10 +720,16 @@ class TestEnvironment < Sprockets::TestCase
     assert_equal 2, asset.metadata[:selector_count]
   end
 
-  test "changing version changes the digest of the asset" do
+  test "changing version changes the digest of the asset (with preprocessor)" do
     old_asset_digest = @env["gallery.js"].hexdigest
     @env.version = 'v2'
     assert old_asset_digest != @env["gallery.js"].hexdigest
+  end
+
+  test "changing version changes the digest of the asset (no preprocessor)" do
+    old_asset_digest = @env["blank.gif"].hexdigest
+    @env.version = 'v2'
+    assert old_asset_digest != @env["blank.gif"].hexdigest
   end
 
   test "bundled asset is stale if its mtime is updated or deleted" do